+++ packages/vue/src/index.ts
@@ -412,4 +412,4 @@ export type { VueImageViewerProps } from './components/ImageViewer'
 export { AspectRatio } from './components/AspectRatio'
 export type { VueAspectRatioProps } from './components/AspectRatio'
 
-export const version = '2.0.19'
+export const version = '2.1.0'